requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pandas==2.2.3
jinja2==3.1.5
python-dotenv==1.0.1
anthropic==0.49.0
